Split index.js file into smaller files
td.replace is the `testdouble` function for mocking a module.
Added `eslint-disable-next-line` due to needing to require
a module dynamically in the test since that is how the module
mocking works
