+++ test/buildTasks.js
@@ -2,7 +2,7 @@
 //
 // Copyright (C) 2016 James Kruth
 //
-// This file is part of nabs.
+// This file is part of
 //
 // nabs is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -15,16 +15,16 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// along with this nabs.  If not, see <http://www.gnu.org/licenses/>.
+// along with this   If not, see <http://www.gnu.org/licenses/>.
 
 'use strict';
 
 const chai = require('chai');
-const nabs = require('../');
+const nabs = require('../src');
 
 chai.should();
 
-describe('buildTasks', () => {
+describe('nabs.buildTasks', () => {
   it('should return a single task if tasks is string/array', () => {
     const name = [];
     let scripts;
